Don't give non-admins read access to /db/_security
This prevents non-admin users from reading a DB's security object.

Previous to this change, it was possible that anonymous users were able
to read a DB's security object if the security object's `members` array
was empty or missing. Also, it was generally possible for authenticated
members (non-admin users) to read the security object, which should not
be possible.